support for locales
As recommended by Deadly I have made forms inside a json file per
locale.

I have also changed this so that no longer will it force append “ - <
form > “ and it will rely on the user to specify where they want the <
form > tag in alarms.json. Trade off is it requires a seperate manager
just for 201 Unown but that’s ok I think?